Change approach for getting cloud name
