+++ src/tree/contexts/aci/AciContextCreateStep.ts
@@ -7,7 +7,6 @@ import { Progress } from 'vscode';
 import { AzureWizardExecuteStep, parseError } from 'vscode-azureextensionui';
 import { ext } from '../../../extensionVariables';
 import { localize } from '../../../localize';
-import { promptForAciCloud } from '../../../utils/azureUtils';
 import { executeAsTask } from '../../../utils/executeAsTask';
 import { execAsync } from '../../../utils/spawnAsync';
 import { IAciWizardContext } from './IAciWizardContext';
@@ -31,7 +30,7 @@ export class AciContextCreateStep extends AzureWizardExecuteStep<IAciWizardConte
             if (error.errorType === '5' || /not logged in/i.test(error.message)) {
                 // If error is due to being not logged in, we'll go through login and try again
                 // Because login could involve device auth we do this step in the terminal
-                await executeAsTask(wizardContext, `docker login azure --cloud-name ${await promptForAciCloud(wizardContext)}`, localize('vscode-docker.commands.contexts.create.aci.azureLogin', 'Azure Login'), { rejectOnError: true });
+                await executeAsTask(wizardContext, `docker login azure --cloud-name ${wizardContext.environment.name}`, localize('vscode-docker.commands.contexts.create.aci.azureLogin', 'Azure Login'), { rejectOnError: true });
                 await execAsync(command);
             } else {
                 // Otherwise rethrow
